enhanced_text_processing metadata code moved over from fetch_and_process.py to hybrid_indexer.py and fetch_and_process.py removed because it's unused
